Renaming db files
