+++ lib/chainsync/chainsync/db/base/__init__.py
@@ -1,7 +1,7 @@
 """Generic database utilities"""
 
 from .conversions import convert_scaled_value_to_decimal
-from .db_interface import (
+from .interface import (
     PostgresConfig,
     TableWithBlockNumber,
     add_user_map,
@@ -14,4 +14,4 @@ from .db_interface import (
     initialize_session,
     query_tables,
 )
-from .db_schema import Base, UserMap
+from .schema import Base, UserMap
